Fix NodeJS module resolution, rename coomposable > composable
