+++ src/composable/use-text-middle-ellipsis.ts
